separate api and lib
