+++ pygld/api/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+
+# Copyright © 2018 Jean-Sébastien Gosselin
+# https://github.com/jnsebgosselin/qwatson
+#
+# This file is part of PyGLD.
+# Licensed under the terms of the GNU General Public License.
+
+"""
+This is an object-oriented plotting library for designing coupled 
+ground-loop heat exchanger.
+"""
