+++ pygld/api/heatpumps/__init__.py
@@ -6,9 +6,8 @@
 # This file is part of PyGLD.
 # Licensed under the terms of the GNU General Public License.
 """
-Everything related to the material properties that are used in the construction
-of a coupled ground-loop heat exchanger system.
+Everything related to the heat pump.
 """
 import os.path as osp
 from pygld import __rootdir__
-__datadir__ = osp.join(__rootdir__, 'materials', 'data')
+__datadir__ = osp.join(__rootdir__, 'api', 'heatpumps', 'data')
