+++ pygld/api/heatpumps/data/TCHV096.hp
