+++ pygld/api/heatpumps/data/TCHV192.hp
