+++ pygld/api/heatpumps/data/TCHV240.hp
