+++ pygld/api/heatpumps/data/TMW360.hp
