+++ pygld/api/heatpumps/data/TMW600.hp
