+++ pygld/api/heatpumps/data/TMW840.hp
