+++ pygld/api/materials/__init__.py
@@ -6,8 +6,9 @@
 # This file is part of PyGLD.
 # Licensed under the terms of the GNU General Public License.
 """
-Everything related to the heat pump.
+Everything related to the material properties that are used in the construction
+of a coupled ground-loop heat exchanger system.
 """
 import os.path as osp
 from pygld import __rootdir__
-__datadir__ = osp.join(__rootdir__, 'heatpumps', 'data')
+__datadir__ = osp.join(__rootdir__, 'api', 'materials', 'data')
