+++ pygld/api/materials/data/proptables_ethylglycol.npy
