+++ pygld/api/materials/data/proptables_propglycol.npy
