+++ pygld/api/materials/data/proptables_purewater.npy
